requirements.txt
streamlit==1.44.1
chromadb==0.5.5
sentence-transformers==3.2.1
langchain==0.3.7
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
google-generativeai==0.8.3
google-genai==1.33.0
python-dotenv==1.0.1
pillow==10.4.0
svgwrite==1.4.3
reportlab==4.2.2
markdown==3.7
PyMuPDF==1.24.13
rapidocr-onnxruntime==1.2.3
pypdfium2==5.7.1
python-pptx==1.0.2
